MULE-5353: Axis outbound connections leak memory
UniversalSender calls muleContext.getRegistry().lookupEndpointFactory().getOutboundEndpoint(uri)
for each message that has to be processed. That calls DefaultTransportServiceDescriptor#createOutboundTransformers
which creates an Instance of ObjectToHttpClientMethodRequest.
After that, the new endpoint id registered in the Registry, that should override any previous instance already
registered. The problem is that we use the hashcode and every instance is returning a different value because
ObjectToHttpClientMethodRequest does not implements hashcode method. That produces the memory leak as the number
of endpoints grows proportional to the number of processed messages.

Fix: adding a naive hashcode implementation to the mentioned transformer.





git-svn-id: https://svn.codehaus.org/mule/branches/mule-2.2.x@21057 bf997673-6b11-0410-b953-e057580c5b09