MULE-5790: Under heavy load and with the IBM JDK transformers are not being applied
The problem is that under heavy load the hashcode generates the same value for two different transformers. And when a list of transformers is checked to see if is was applied or not it returns the wrong value, resulting in the transformers being skipped. As a workaround I added the name of the transformer to the generation of the list hashcode, since overriding the Transformer hashcode would have greater impact and this is a hack for 2.2.x to ensure that a list of transformers is not being applied twice

git-svn-id: https://svn.codehaus.org/mule/branches/mule-2.2.x@23314 bf997673-6b11-0410-b953-e057580c5b09